Move sanitizePII unit tests into sanitize.test.js file (PR #1986)
